Move connection cache and HostSession from init to command
Move the session handling to command.py. This puts things in one
location and should make testing easier.
